+++ htab_for_each.c
@@ -3,7 +3,7 @@
 
 
 void htab_for_each(const htab_t * t, void (*f)(htab_pair_t *data)){
-    for(int i = 0; i < t->arr_size; i++){
+    for(size_t i = 0; i < t->arr_size; i++){
         htab_itm_t *itm = t->arr_ptr[i];
         while(itm != NULL) {
             f(&itm->pair);
